+++ AuthPermissions.AspNetCore/StartupServices/StartupServiceMigrateAnyDbContext.cs
@@ -5,23 +5,27 @@ using System;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
+using RunMethodsSequentially;
 
-namespace AuthPermissions.AspNetCore.InternalStartupServices
+namespace AuthPermissions.AspNetCore.StartupServices
 {
     /// <summary>
     /// This will run EF Core's Migate method on the given DbContext
     /// Note that if the database is an in-memory, then it will simply create it
     /// </summary>
-    internal class StartupMigrateAnyDbContext<TContext> where TContext : DbContext
+    public class StartupServiceMigrateAnyDbContext<TContext> : IStartupServiceToRunSequentially 
+        where TContext : DbContext
     {
+        public int OrderNum { get; } = -10; //These must be run before any other startup services
+
         /// <summary>
         /// This migrates the given <typeparamref name="TContext"/> DbContext
         /// </summary>
-        /// <param name="scopedService">This should be a scoped service</param>
+        /// <param name="scopedServices">This should be a scoped service</param>
         /// <returns></returns>
-        public async ValueTask StartupServiceAsync(IServiceProvider scopedService)
+        public async ValueTask ApplyYourChangeAsync(IServiceProvider scopedServices)
         {
-            var context = scopedService.GetRequiredService<TContext>();
+            var context = scopedServices.GetRequiredService<TContext>();
 
             if (context.Database.IsInMemory())
                 await context.Database.EnsureCreatedAsync();