+++ AuthPermissions.AspNetCore/StartupServices/StartupServicesIndividualAccountsAddDemoUsers.cs
@@ -8,22 +8,27 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using RunMethodsSequentially;
 
-namespace AuthPermissions.AspNetCore.InternalStartupServices
+namespace AuthPermissions.AspNetCore.StartupServices
 {
-    internal class StartupIndividualAccountsAddDemoUsers
+    public class StartupServicesIndividualAccountsAddDemoUsers : IStartupServiceToRunSequentially
     {
+        //This must be run after the migation of the IndividualAccounts database
+        //But before the SuperUser is added
+        public int OrderNum { get; } = -5; 
+
         /// <summary>
         /// This takes a comma delimited string of demo users from the "DemoUsers" in the appsettings.json file
         /// and adds each if they aren't in the individual account user
         /// NOTE: The email is also the password, so make sure the email is a valid password
         /// </summary>
-        /// <param name="scopedService">This should be a scoped service</param>
+        /// <param name="scopedServices">This should be a scoped service</param>
         /// <returns></returns>
-        public async ValueTask StartupServiceAsync(IServiceProvider scopedService)
+        public async ValueTask ApplyYourChangeAsync(IServiceProvider scopedServices)
         {     
-            var userManager = scopedService.GetRequiredService<UserManager<IdentityUser>>();
-            var config = scopedService.GetRequiredService<IConfiguration>();
+            var userManager = scopedServices.GetRequiredService<UserManager<IdentityUser>>();
+            var config = scopedServices.GetRequiredService<IConfiguration>();
             var demoUsers = config["DemoUsers"];
 
             if (!string.IsNullOrEmpty(demoUsers))