+++ Example3.InvoiceCode/AppStart/SeedInvoiceDbContext.cs
@@ -1,14 +1,11 @@
 ﻿// Copyright (c) 2021 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
 // Licensed under MIT license. See License.txt in the project root for license information.
 
-using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using AuthPermissions.AdminCode;
 using AuthPermissions.DataLayer.Classes;
 using Example3.InvoiceCode.EfCoreClasses;
 using Example3.InvoiceCode.EfCoreCode;
-using Microsoft.EntityFrameworkCore;
 
 namespace Example3.InvoiceCode.AppStart
 {